Remove uint16 cast for CDS images for JWSTFGS-266

+++ jwst_magic/fsw_file_writer/write_files.py
@@ -210,7 +210,7 @@ def write_cds(obj):
         filename_cds = os.path.join(obj.out_dir,
                                     obj.stsci_dir,
                                     obj.filename_root + 'cds.fits')
-        utils.write_fits(filename_cds, np.uint16(obj.cds), log=LOGGER)
+        utils.write_fits(filename_cds, obj.cds, log=LOGGER)
 
 
 def write_image(obj):
